+++ rust/candid_parser/build.rs
@@ -1,8 +1,7 @@
 fn main() {
-    #[cfg(feature = "parser")]
     lalrpop::Configuration::new()
         .use_cargo_dir_conventions()
         .emit_rerun_directives(true)
-        .process_file("src/parser/grammar.lalrpop")
+        .process_file("src/grammar.lalrpop")
         .unwrap();
 }